# OPENAPI_SWAGGER_SPEC.yaml
openapi: 3.0.0
info:
  title: Asset Management Enterprise SaaS API
  description: Complete API documentation for Asset Management system with enterprise features
  version: 1.0.0
  contact:
    name: API Support
    url: https://example.com/support
    email: api-support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8085/api/v1
    description: Local Development
  - url: https://api.example.com/api/v1
    description: Production
paths:
...
  /assets/{assetId}:
    get:
      summary: Get asset by ID
      operationId: getAsset
      tags:
        - Assets
      security:
        - bearerAuth: []
      parameters:
        - name: assetId
          in: path
          required: true
          description: Asset UUID
          schema:
            type: string
      responses:
        '200':
          description: Asset details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '404':
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update asset
      operationId: updateAsset
      tags:
        - Assets
      security:
        - bearerAuth: []
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetRequest'
      responses:
        '200':
          description: Asset updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '404':
          description: Asset not found
        '400':
          description: Validation error

    delete:
      summary: Delete asset
      operationId: deleteAsset
      tags:
        - Assets
      security:
        - bearerAuth: []
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Asset deleted
        '404':
          description: Asset not found
...
  /cloud-assets/sync:
    post:
      summary: Trigger cloud asset discovery for one provider
      operationId: syncCloudAssets
      tags:
        - Cloud Assets
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - provider
              properties:
                provider:
                  type: string
                  enum: [AWS, AZURE, GCP]
                  description: Cloud provider to scan
                regions:
                  type: array
                  items:
                    type: string
                  description: Provider-specific region codes. Omit to use the configured default.
                  example: ["us-east-1", "eu-west-1"]
      responses:
        '200':
          description: Sync completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncResult'
        '400':
          description: Missing or invalid provider
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: "Unknown provider 'XYZ'. Valid values: AWS, AZURE, GCP."
        '401':
          description: Not authenticated
        '403':
          description: Insufficient authority
  /cloud-assets/sync/all:
    post:
      summary: Trigger cloud asset discovery across all configured providers
      operationId: syncAllCloudAssets
      tags:
        - Cloud Assets
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                regions:
                  type: array
                  items:
                    type: string
                  description: Region codes forwarded to all providers. Omit to use provider defaults.
                  example: ["us-east-1", "europe-west1"]
      responses:
        '200':
          description: All-provider sync completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncResult'
        '401':
          description: Not authenticated
        '403':
          description: Insufficient authority

  # ── BOG Compliance ────────────────────────────────────────────────────────────
...
  /checkouts/assets/{assetId}:
    get:
      summary: List checkouts for asset
      operationId: listCheckoutsByAsset
      tags:
          - Checkouts
      security:
        - BearerAuth: []
      parameters:
        - name: assetId
          in: path
          required: true
          description: Asset UUID
          schema:
            type: string
      responses:
        '200':
          description: Checkouts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CheckoutRecordDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /cloud-assets:
    get:
      summary: List cloud assets
      operationId: listCloudAssets
      tags:
          - Cloud Assets
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Cloud assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CloudAssetDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /cloud-assets/cost-summary:
    get:
      summary: Aggregated cloud cost summary
      operationId: getCloudCostSummary
      tags:
          - Cloud Assets
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Cost summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalMonthlyCost:
                    type: number
                  byProvider:
                    type: object
                  byService:
                    type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /cloud-assets/{id}:
    delete:
      summary: Delete a cloud asset
      operationId: deleteCloudAsset
      tags:
          - Cloud Assets
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Cloud asset UUID
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
  /cloud-assets/{id}/cost:
    post:
      summary: Record cost entry for cloud asset
      operationId: recordCloudCost
      tags:
          - Cloud Assets
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Cloud asset UUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                currency:
                  type: string
                period:
                  type: string
                  description: YYYY-MM
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudAssetDto'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
...
  /leases/assets/{assetId}:
    get:
      summary: List leases for an asset
      operationId: listLeasesByAsset
      tags:
          - Leases
      security:
        - BearerAuth: []
      parameters:
        - name: assetId
          in: path
          required: true
          description: Asset UUID
          schema:
            type: string
      responses:
        '200':
          description: Leases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaseRecordDto'
        '401':
          description: Unauthorized
        '404':
          description: Not found
...
  /compliance/controls/{id}:
    delete:
      summary: Delete Control
      operationId: deleteCompliance_controls
      tags:
          - Compliance
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Control UUID
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AssetRequest:
      type: object
      required:
        - name
        - assetTag
        - categoryId
        - departmentId
        - purchaseCost
        - currency
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        assetTag:
          type: string
        assetType:
          type: string
        serialNumber:
          type: string
        categoryId:
          type: string
          format: uuid
        departmentId:
          type: string
          format: uuid
        locationId:
          type: string
          format: uuid
        supplierId:
          type: string
          format: uuid
        purchaseCost:
          type: number
          format: double
        currency:
          type: string
          enum: [USD, EUR, GBP]
        status:
          type: string
          enum: [IN_STOCK, IN_USE, RETIRED]
        condition:
          type: string
          enum: [POOR, FAIR, GOOD, EXCELLENT]
    Asset:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        assetTag:
          type: string
        status:
          type: string
        condition:
          type: string
        purchaseCost:
          type: number
        currentBookValue:
          type: number
        organisationId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    SyncResult:
      type: object
      properties:
        status:
          type: string
          example: completed
        provider:
          type: string
          example: AWS
          description: Present on single-provider sync; absent on syncAll
        assetsUpserted:
          type: integer
          example: 42
        message:
          type: string
          example: "AWS asset sync completed. 42 asset(s) upserted."
...
    Error:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...